Renamed wrr to weightedroundrobin Reverted balancer/base/balancer.go Renamed wrr.Info to weightedroundrobin.AddrInfo
