+++ balancer/base/balancer.go
@@ -20,7 +20,6 @@ package base
 
 import (
 	"context"
-	"google.golang.org/grpc/balancer/wrr"
 
 	"google.golang.org/grpc/balancer"
 	"google.golang.org/grpc/connectivity"
@@ -39,7 +38,7 @@ func (bb *baseBuilder) Build(cc balancer.ClientConn, opt balancer.BuildOptions) 
 		cc:            cc,
 		pickerBuilder: bb.pickerBuilder,
 
-		subConns: make(map[resolver.Address]AddrInfo),
+		subConns: make(map[resolver.Address]balancer.SubConn),
 		scStates: make(map[balancer.SubConn]connectivity.State),
 		csEvltr:  &balancer.ConnectivityStateEvaluator{},
 		// Initialize picker to a picker that always return
@@ -61,7 +60,7 @@ type baseBalancer struct {
 	csEvltr *balancer.ConnectivityStateEvaluator
 	state   connectivity.State
 
-	subConns map[resolver.Address]AddrInfo
+	subConns map[resolver.Address]balancer.SubConn
 	scStates map[balancer.SubConn]connectivity.State
 	picker   balancer.Picker
 	config   Config
@@ -86,11 +85,7 @@ func (b *baseBalancer) UpdateClientConnState(s balancer.ClientConnState) {
 				grpclog.Warningf("base.baseBalancer: failed to create new SubConn: %v", err)
 				continue
 			}
-			addrInfo := AddrInfo{SubConn: sc}
-			if wrrInfo, ok := a.Metadata.(*wrr.Info); ok {
-				addrInfo.Weight = wrrInfo.Weight
-			}
-			b.subConns[a] = addrInfo
+			b.subConns[a] = sc
 			b.scStates[sc] = connectivity.Idle
 			sc.Connect()
 		}
@@ -98,7 +93,7 @@ func (b *baseBalancer) UpdateClientConnState(s balancer.ClientConnState) {
 	for a, sc := range b.subConns {
 		// a was removed by resolver.
 		if _, ok := addrsSet[a]; !ok {
-			b.cc.RemoveSubConn(sc.SubConn)
+			b.cc.RemoveSubConn(sc)
 			delete(b.subConns, a)
 			// Keep the state of this sc in b.scStates until sc's state becomes Shutdown.
 			// The entry will be deleted in HandleSubConnStateChange.
@@ -115,28 +110,15 @@ func (b *baseBalancer) regeneratePicker() {
 		b.picker = NewErrPicker(balancer.ErrTransientFailure)
 		return
 	}
-	pickerBuilderV2, ok := b.pickerBuilder.(PickerBuilderV2)
-	if ok {
-		readySCs := make(map[resolver.Address]AddrInfo)
+	readySCs := make(map[resolver.Address]balancer.SubConn)
 
-		// Filter out all ready SCs from full subConn map.
-		for addr, sc := range b.subConns {
-			if st, ok := b.scStates[sc.SubConn]; ok && st == connectivity.Ready {
-				readySCs[addr] = sc
-			}
+	// Filter out all ready SCs from full subConn map.
+	for addr, sc := range b.subConns {
+		if st, ok := b.scStates[sc]; ok && st == connectivity.Ready {
+			readySCs[addr] = sc
 		}
-		b.picker = pickerBuilderV2.BuildV2(readySCs)
-	} else {
-		readySCs := make(map[resolver.Address]balancer.SubConn)
-
-		// Filter out all ready SCs from full subConn map.
-		for addr, sc := range b.subConns {
-			if st, ok := b.scStates[sc.SubConn]; ok && st == connectivity.Ready {
-				readySCs[addr] = sc.SubConn
-			}
-		}
-		b.picker = b.pickerBuilder.Build(readySCs)
 	}
+	b.picker = b.pickerBuilder.Build(readySCs)
 }
 
 func (b *baseBalancer) HandleSubConnStateChange(sc balancer.SubConn, s connectivity.State) {
