+++ balancer/weightedroundrobin/weightedroundrobin.go
@@ -1,9 +1,9 @@
-package wrr
+package weightedroundrobin
 
 // Name is the name of weighted_round_robin balancer.
 const Name = "weighted_round_robin"
 
 // Information that should be stored inside Address metadata in order to use wrr.
-type Info struct{
+type AddrInfo struct{
 	Weight uint32
 }
